+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add ability to use cursor-based pagination instead of offset-based pagination.
   - Cursor-based pagination can be 30-50% faster than offset-based pagination for large datasets.
   - Enable cursor pagination by passing `use_cursor_pagination=True` to `client.get_dataset()`.
+- Makes cursor-based pagination the default for `client.get_dataset()`.
 
 ## v0.5.9 - March 26, 2024